Create Download_Jars.sh
The last few jars are really big and only need to be downloaded once to your github repo. So as not to take up space, I have included a link to the zip file in my public dropbox, it should be live in about an hour from the time of this upload.